Removed employment history section, added side projects section
If employers are interested in employment history they would be better served viewing my full linkedin profile available through the icon on the home page.

Side project section showcases recent projects that I've been working on with a screenshot that links to the github repository in question. Some information and context added as a description underneath the image too.